+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-_No unreleased changes yet._
+### Added
+- Made the expanded queue pane a clear drop target for single and multi-item media drags
 
 ## [2.4.0] - 2026-09-10